[DOK-225] style : Admincomponents 나눔
