+++ src/components/atoms/CKEditor.js
@@ -1,0 +1,24 @@
+import React, { useState } from "react";
+import CKEditor from "@ckeditor/ckeditor5-react";
+import ClassicEditor from "@ckeditor/ckeditor5-build-classic";
+
+
+// CKEditor
+export const CKEditorComponent = () => {
+  const [editorData, setEditorData] = useState("");
+
+  const handleEditorChange = (event, editor) => {
+    const data = editor.getData();
+    setEditorData(data);
+  };
+
+  return (
+    <div>
+      <CKEditor
+        editor={ClassicEditor}
+        data={editorData}
+        onChange={handleEditorChange}
+      />
+    </div>
+  );
+};
