+++ src/components/atoms/PageNation.js
@@ -1,0 +1,6 @@
+import { Pagination } from "@mui/material";
+
+// 페이지네이션
+export default function BasicPagination() {
+    return <Pagination count={10} />;
+  }
