+++ src/pages/Admin/CSManage/FAQPage.js
@@ -6,7 +6,7 @@ import {
   InputBoxS,
   InputBoxM,
   CustomSelect,
-} from "../../../components/atoms/AdminComponents";
+} from "../../../components/atoms/CKEditor";
 import SearchIcon from "@mui/icons-material/Search";
 import InfoOutlinedIcon from "@mui/icons-material/InfoOutlined";
 import { CKEditor } from "@ckeditor/ckeditor5-react";
